[refactor] Refactor Navbar component - Created new `SearchBar` component to handle search functionality. - Created new `AuthControls` component to manage login/logout. - Updated `Navbar.jsx` to use `SearchBar` and `AuthControls`. - Improved code organization and readability for better maintainability and testability.
