+++ src/components/Navbar/Navbar.jsx
@@ -1,77 +1,36 @@
+import ThemeIcon from '@/assets/icons/theme.svg';
+import MenuButton from '@/components/Buttons';
+import { menu } from '@/data/menu';
+import { toggleThemeChange } from '@/store/themeSlice';
+import { Container } from '@/styles';
 import { useState } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { useNavigate } from 'react-router-dom';
+import { AuthControls } from './AuthControls';
 import {
-  FormGroup,
+  BuggerIcon,
   Header,
   HeaderContainer,
-  Input,
-  MenuWrapper,
   Image,
-  SearchIcon,
-  Login,
-  Logout,
-  Theme,
   Label,
   MenuThemeItem,
+  MenuWrapper,
   StyledBars3Icon,
-  BuggerIcon,
   StyledXMarkIcon,
+  Theme,
 } from './Navbar.styles';
-import { Container } from '@/styles';
-import ThemeIcon from '@/assets/icons/theme.svg';
-import { menu } from '@/data/menu';
-import MenuButton from '@/components/Buttons';
-import { toggleThemeChange } from '@/store/themeSlice';
-import { getAuth, signOut } from '@firebase/auth';
-import { signInWithGoogle } from '@/firebase/firebaseAuth';
+import { SearchBar } from './SearchBar';
 
 const Navbar = () => {
-  const dispatch = useDispatch();
-  const navigate = useNavigate();
-  const [query, setQuery] = useState('');
   const [isOpen, setIsOpen] = useState(false);
-
-  const user = useSelector((state) => state.auth.user);
 
   const handleToggle = () => {
     dispatch(toggleThemeChange());
-  };
-
-  const handleSearch = (e) => {
-    setQuery(e.target.value);
-  };
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    navigate(`/search/photos/${query}`);
-    setQuery('');
-  };
-
-  const handleLogout = async () => {
-    const auth = getAuth();
-
-    await signOut(auth).catch((error) => {
-      console.error('Firebase sign-out error:', error);
-    });
   };
 
   return (
     <Header>
       <Container>
         <HeaderContainer>
-          <form onSubmit={handleSubmit}>
-            <FormGroup>
-              <SearchIcon />
-              <Input
-                value={query}
-                onChange={handleSearch}
-                type='text'
-                placeholder='Search...'
-              />
-            </FormGroup>
-          </form>
-
+          <SearchBar />
           <BuggerIcon>
             {isOpen ? (
               <StyledXMarkIcon onClick={() => setIsOpen(!isOpen)} />
@@ -86,17 +45,12 @@ const Navbar = () => {
               ))}
               <MenuThemeItem onClick={handleToggle}>
                 <Theme>
-                  <Image src={ThemeIcon} alt='Theme Icon' />
+                  <Image src={ThemeIcon} alt="Theme Icon" />
                 </Theme>
                 <Label>Theme</Label>
               </MenuThemeItem>
-              {/* check if user is authenticated */}
-              {!user && (
-                <Login onClick={() => dispatch(signInWithGoogle())}>
-                  Login
-                </Login>
-              )}
-              {user && <Logout onClick={handleLogout}>Log out</Logout>}
+
+              <AuthControls />
             </>
           </MenuWrapper>
         </HeaderContainer>
